fixup! perf(ivy): cache multiple reads to an element's stylingContext
