+++ packages/core/src/render3/state.ts
@@ -12,8 +12,8 @@ import {assertLViewOrUndefined} from './assert';
 import {executeHooks} from './hooks';
 import {ComponentDef, DirectiveDef} from './interfaces/definition';
 import {TElementNode, TNode, TViewNode} from './interfaces/node';
-import {StylingContext} from './interfaces/styling';
 import {BINDING_INDEX, CONTEXT, DECLARATION_VIEW, FLAGS, InitPhaseState, LView, LViewFlags, OpaqueViewState, TVIEW} from './interfaces/view';
+import {setCachedStylingContext} from './styling/state';
 import {resetPreOrderHookFlags} from './util/view_utils';
 
 
@@ -523,26 +523,4 @@ export function ɵɵnamespaceHTML() {
 
 export function getNamespace(): string|null {
   return _currentNamespace;
-}
-
-let stylingContext: StylingContext|null = null;
-
-/**
- * Gets the most recent styling context value.
- *
- * Note that only one styling context is stored at a given time.
- */
-export function getCachedStylingContext() {
-  return stylingContext;
-}
-
-/**
- * Sets the most recent styling context value.
- *
- * Note that only one styling context is stored at a given time.
- *
- * @param context The styling context value that will be stored
- */
-export function setCachedStylingContext(context: StylingContext | null) {
-  stylingContext = context;
 }
